# src/main/resources/META-INF/openapi.yaml
---
openapi: 3.0.3
info:
  title: Generated API
  version: "1.0"
paths:
  /job-offers:
    get:
      parameters:
      - name: page
        in: query
        schema:
          format: int32
          default: 0
          type: integer
      - name: size
        in: query
        schema:
          format: int32
          default: 20
          type: integer
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobOffer'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobOffer'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: ID of the created offer.
  /job-offers/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOffer'
  /people:
    get:
      parameters:
      - name: page
        in: query
        schema:
          format: int32
          default: 0
          type: integer
      - name: size
        in: query
        schema:
          format: int32
          default: 20
          type: integer
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRegistration'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: ID of the created person.
        "400":
          description: "Bad request"

  /people/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        "404":
          description: Not Found
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdate'
      responses:
        "204":
          description: OK
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: OK
components:
  schemas:
    JobOffer:
      type: object
      properties:
        category:
          type: string
          enum: [IT, FoodDrinks,Office,Courier,ShopAssistant]
        employerId:
          type: string
        endDate:
          format: date
          type: string
        id:
          format: int64
          type: integer
        startDate:
          format: date
          type: string
    Person:
      type: object
      properties:
        creationDate:
          format: date
          type: string
        id:
          format: int64
          type: integer
        login:
          type: string
        name:
          type: string
    PersonRegistration:
      type: object
      properties:
        login:
          type: string
        name:
          type: string
        password:
          type: string
      required: 
        - login
        - name
        - password
    PersonUpdate:
      type: object
      properties:
        name:
          type: string
        password:
          type: string